Added branch detection and interactive user confirmation for branching.

+++ lib/kenny_patch.rb
@@ -8,34 +8,43 @@ class KennyPatch
 
     # create patch directory
     @patch_dir = @repo.commits_path + File::Separator + @uuid
+    Dir.mkdir(@patch_dir) if !File.exist?(@patch_dir)
 
     # create parent directory
     @parent_dir = @patch_dir + File::Separator + @repo.get_current
 
-    # make an empty children file
-    File.open(@patch_dir + File::Separator + 'children','w') {|f| }
-    
     # check existence
     if fname.nil? && File.exist?(@patch_dir) && File.directory?(@patch_dir)
       @fname = File.read(@patch_dir + File::Separator + "file").strip
     end
   end
-  
+
   def process_fname(name)
     # Do nothing - no common behavior
   end
-  
+
   def type
     "unknown"
   end
 
   def create
-    puts 'Creating ' + @patch_dir + '.'
-    Dir.mkdir(@patch_dir)
-    
+    parent_children = @repo.commits_path + File::Separator + @repo.get_current + File::Separator + 'children'
+
+    if File.exist?(parent_children)
+      if File.size(parent_children) > 0
+        puts "This patch already has children. Are you sure you want to create a branch? (y/n)"
+        if $STDIN.gets.chomp != 'y'
+          exit
+        end
+      end
+    end
+
+    puts 'Creating children file.'
+    File.open(@patch_dir + File::Separator + 'children','w') {|f| }
+
     puts "Recording type #{self.type}."
     File.open(@patch_dir + File::Separator + "type", "w") {|f| f << self.type}
-    
+
     puts "Recording file name #{@fname}."
     File.open(@patch_dir + File::Separator + "file", "w") {|f| f << @fname}
 
@@ -43,7 +52,7 @@ class KennyPatch
     Dir.mkdir(@parent_dir)
 
     # add new patch as a child to the current patch
-    File.open(@repo.commits_path + File::Separator + @repo.get_current + File::Separator + 'children', 'a') << @uuid + "\n"
+    File.open(parent_children,'w') << @uuid + "\n"
 
     # return uuid
     @uuid
@@ -62,7 +71,7 @@ class KennyPatch
       raise 'Patch ' + @uuid + ' does not exist.'
     end
   end
-  
+
   def parents
     parent_list = []
     Dir.new(@patch_dir).each do |entry|
@@ -76,11 +85,11 @@ class KennyPatch
 
   def children
     children_list = []
-    File.open(@patch_dir + File::Separator + 'children').each { |line| 
+    File.open(@patch_dir + File::Separator + 'children').each { |line|
       children_list.push(@repo.patch_for_uuid(line.strip)) }
     return children_list
   end
-  
+
   def find_ancestry(target_fname)
     # Find one possible path from this patch to the given file's creation
     ancestry = self.parents[0].find_ancestry(target_fname)
